refactor: implement separate custom fetch data hook for posts
