+++ src/app/store/page.js
@@ -7,7 +7,7 @@ import Image from "next/image";
 import "./Store.css";
 import "../mediaQueries.css";
 import useProductsUrl from "../hooks/useProductsUrl";
-import useFetchData from "../hooks/useFetchData";
+// import useFetchData from "../hooks/useFetchPosts";
 import DeleteButton from "../components/DeleteButton";
 
 export default function Store({ searchParams }) {
@@ -23,24 +23,26 @@ export default function Store({ searchParams }) {
 
   const productsUrl = useProductsUrl(searchQuery, sortOptions, filter);
 
-  const { list, setList } = useFetchData("products", productsUrl);
+  // const { list, setList } = useFetchData("products", productsUrl);
 
   function handleDelete(id) {
-    localStorage.setItem(
-      "products",
-      JSON.stringify(list.filter((product) => product.id !== id))
-    );
-    setList((curProducts) =>
-      curProducts.filter((product) => product.id !== id)
-    );
+    // localStorage.setItem(
+    //   "products",
+    //   JSON.stringify(list.filter((product) => product.id !== id))
+    // );
+    // setList((curProducts) =>
+    //   curProducts.filter((product) => product.id !== id)
+    // );
   }
+
+  const products = [];
 
   return (
     <section className="product__page-wrapper">
       <h1 className="section__header">Products</h1>
       <div className="product__page-content">
         <ProductFilter />
-        <ProductList products={list} onDelete={handleDelete} />
+        <ProductList products={products} onDelete={handleDelete} />
       </div>
     </section>
   );
